fixed naming and added file errors.py.
